+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [unreleased]
+
+- Add `String` number type to disable numeric conversions
+
 ## [0.1.1]
 
 - Improve handling of whitespace in shared strings